# BACKEND/pic/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: PIC DIAN
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of the responses to the client (Accepts)
produces:
  - application/json
  - application/xml
paths:
...
  /cargarUsuariosExcel:
      x-swagger-router-controller: user_j_controller
      post:
        description: Carga de usuarios de un excel
        operationId: CargarEmpleado
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /buscarEmpleado:
      x-swagger-router-controller: user_j_controller
      post:
        description: buscarUsuario
        operationId: buscarEmpleado
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
  /listaNiveles:
      x-swagger-router-controller: user_j_controller
      post:
        description: listado de niveles
        operationId: listaNiveles
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
...
  /agregarCuestionario:
      x-swagger-router-controller: cuestionario.controller
      post:
        description: agregar cuestionario
        operationId: Cuestionario
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
...
  /cambioPassword:
      x-swagger-router-controller: user_j_controller
      post:
        description: Cambio Contraseña
        operationId: cambioPassword
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
...
  /crearNuevoUsuario:
      x-swagger-router-controller: user_j_controller
      post:
        description: crear Nuevo usuario
        operationId: crearNuevoUsuario
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
...
  /reporte_UsuarioC:
      x-swagger-router-controller: reportes_controller
      post:
        description: excel del reporte de usuario y cuestionario
        operationId: estado_Cuestionario
        responses:
          "200":
            description: Success
            schema:
              $ref: "#/definitions/GenericResponse"
          default:
            description: Error
            schema:
              $ref: "#/definitions/ErrorResponse"
...
definitions:
  GenericResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string